+++ app/hooks/useProfileManager.ts
@@ -1,0 +1,63 @@
+import { useState } from "react";
+import { useAuth } from "../contexts/AuthContext.web";
+import { useProfile } from "./useProfile";
+import { UpdateProfileRequest } from "../utils/api/profile";
+
+/**
+ * プロフィール管理のカスタムフック
+ * プロフィールの取得、更新、編集シートの状態管理を一元化する
+ */
+export const useProfileManager = () => {
+  const { user, logout, isAuthenticated, getAccessToken } = useAuth();
+  const [isEditSheetOpen, setIsEditSheetOpen] = useState(false);
+
+  // ユーザーIDを取得（Auth0のsubフィールドから）
+  const userId = user?.sub;
+
+  // プロフィール情報を取得・管理
+  const { profile, isLoading: isProfileLoading, updateProfile } = useProfile(
+    userId,
+    getAccessToken,
+    user ? {
+      email: user.email,
+      name: user.name,
+      picture: user.picture,
+    } : undefined
+  );
+
+  const handleProfileUpdate = async (updateData: UpdateProfileRequest) => {
+    await updateProfile(updateData);
+  };
+
+  const handleEditButtonClick = () => {
+    if (profile) {
+      setIsEditSheetOpen(true);
+    }
+    // プロフィール情報がまだ読み込まれていない場合は何もしない（ボタンがdisabledになる）
+  };
+
+  const handleEditSheetClose = () => {
+    setIsEditSheetOpen(false);
+  };
+
+  return {
+    // ユーザー情報
+    user,
+    logout,
+    isAuthenticated,
+    userId,
+    
+    // プロフィール情報
+    profile,
+    isProfileLoading,
+    
+    // 編集シート状態
+    isEditSheetOpen,
+    setIsEditSheetOpen,
+    
+    // ハンドラー関数
+    handleProfileUpdate,
+    handleEditButtonClick,
+    handleEditSheetClose,
+  };
+};
